Update expr.cpp
allowing to store expresion result to a named variable.
for example we write on comment constraint `D1 = 10`, after that we can use `D1` on other constraint for example on dimension constraint `D1+0.2` will result 10.2

the variable stored is local variable. and will lost when program close and reopen the file. I still havent find way to re-evalueate all comment constraint to get the named variable back to the local memory.